Add package for shared code, rmw_connext_shared_cpp
