+++ rmw_connext_cpp/package.xml
@@ -13,16 +13,19 @@
   <buildtool_export_depend>rosidl_cmake</buildtool_export_depend>
 
   <build_depend>connext_cmake_module</build_depend>
-  <build_depend>libndds51-dev</build_depend>
+  <build_depend>libndds52-dev</build_depend>
   <build_depend>rmw</build_depend>
+  <build_depend>rmw_connext_shared_cpp</build_depend>
   <build_depend>rosidl_generator_cpp</build_depend>
   <build_depend>rosidl_typesupport_connext_cpp</build_depend>
   <build_depend>rosidl_generator_dds_idl</build_depend>
 
   <build_export_depend>connext_cmake_module</build_export_depend>
-  <build_export_depend>libndds51</build_export_depend>
+  <build_export_depend>libndds52</build_export_depend>
   <build_export_depend>rosidl_generator_cpp</build_export_depend>
   <build_export_depend>rosidl_typesupport_connext_cpp</build_export_depend>
+
+  <exec_depend>rmw_connext_shared_cpp</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
