+++ rmw_connext_dynamic_cpp/package.xml
@@ -11,6 +11,7 @@
   <build_depend>connext_cmake_module</build_depend>
   <build_depend>libndds51-dev</build_depend>
   <build_depend>rmw</build_depend>
+  <build_depend>rmw_connext_shared_cpp</build_depend>
   <build_depend>rosidl_generator_cpp</build_depend>
   <build_depend>rosidl_typesupport_introspection_cpp</build_depend>
