+++ rmw_connext_shared_cpp/package.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>rmw_connext_shared_cpp</name>
+  <version>0.0.0</version>
+  <description>C++ types and functions shared by the ROS middleware interface to RTI Connext Static and RTI Connext Dynamic.</description>
+  <maintainer email="jackie@osrfoundation.org">Jackie Kay</maintainer>
+  <license>Apache License 2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>rosidl_cmake</buildtool_depend>
+
+  <buildtool_export_depend>ament_cmake</buildtool_export_depend>
+
+  <build_depend>connext_cmake_module</build_depend>
+  <build_depend>libndds52-dev</build_depend>
+  <build_depend>rmw</build_depend>
+
+  <build_export_depend>connext_cmake_module</build_export_depend>
+  <build_export_depend>libndds52</build_export_depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
